made sprites take u32 instead of f32 (super breaking change)

+++ src/graphics/sprite.rs
@@ -7,13 +7,13 @@ pub struct Sprite {
 const INSET_FOR_SPRITE_BLEED_FIX: f32 = 1.0 / 100.0;
 
 impl Sprite {
-    pub const fn new(x: f32, y: f32, width: f32, height: f32) -> Self {
+    pub const fn new(x: u32, y: u32, width: u32, height: u32) -> Self {
         Sprite {
             src_rect: Rectangle {
-                x: x + INSET_FOR_SPRITE_BLEED_FIX,
-                y: y + INSET_FOR_SPRITE_BLEED_FIX,
-                width: width - (INSET_FOR_SPRITE_BLEED_FIX),
-                height: height - (INSET_FOR_SPRITE_BLEED_FIX),
+                x: x as f32 + INSET_FOR_SPRITE_BLEED_FIX,
+                y: y as f32 + INSET_FOR_SPRITE_BLEED_FIX,
+                width: width as f32 - (INSET_FOR_SPRITE_BLEED_FIX),
+                height: height as f32 - (INSET_FOR_SPRITE_BLEED_FIX),
             },
         }
     }
